added routes for the matches controllers
