+++ backend/src/routes/api.routes.ts
@@ -1,0 +1,25 @@
+import { Router } from 'express';
+import { getMatches, getMatchById, createMatch, updateMatchStatus, getMatchLog } from '../controllers/matches.controllers';
+import {
+  getUserElo,
+  getEloHistory,
+  updateEloAfterMatch,
+  getLeaderboard
+} from '../controllers/elo.controllers';
+
+const router = Router();
+
+// Match routes
+router.get('/matches', getMatches);
+router.get('/matches/:match_id', getMatchById);
+router.post('/matches', createMatch);
+router.patch('/matches/:match_id/status', updateMatchStatus);
+router.get('/matches/:match_id/log', getMatchLog);
+
+//elo routes
+router.get('/:user_id', getUserElo);
+router.get('/:user_id/history', getEloHistory);
+router.post('/update', updateEloAfterMatch);
+router.get('/leaderboard', getLeaderboard);
+
+export default router;
